Extract functions related to showing the bootstrap status in their own module
